Move a bunch of non-theme client stuff out
Assets are pretty heavily used by the theme client
but that code doesn't need to be in there. Moving it
out also helps since the load_asset_test was basically
testing Assets. Running this test on it's own shouldn't
be weird anymore
